+++ hazelcast/src/test/java/com/hazelcast/internal/dynamicconfig/DynamicConfigYamlGeneratorTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.config.dynamic;
+package com.hazelcast.internal.dynamicconfig;
 
 import com.hazelcast.config.Config;
 import com.hazelcast.config.InMemoryYamlConfig;
